Extract server core helpers
Move slug validation, error formatting, insights batching, and generate-input resolution behind a tested core module while preserving server exports for callers.
